feat: Mise à jour des types de données pour les quantités et les montants, amélioration de la gestion des fichiers et des messages d'erreur

+++ lib/domain/models/order_model.dart
@@ -155,7 +155,7 @@ class OrderSendModel {
   String? giftFrom;
   int? idUrssaf;
   String? giftReason;
-  int? fundQuantity;
+  double? fundQuantity;
   String? fundExpiryDate;
   List<OrderSendItems>? orderItems;
   int? idEntity;
@@ -203,7 +203,7 @@ class OrderSendModel {
 }
 
 class OrderSendItems {
-  int? amount;
+  double? amount;
   int? quantity;
   String? persoMsg;
